rt_ioctl_iwaplist: fix memcpy
Kernel memcpy checks detected an issue with the use of &addr. The
memcpy was incorrect: &addr should be addr. This error seems to be the
result of addr being changed from an array to a pointer at some stage,
since &addr and addr are the same value if the type is an array.